Added Viewport, Undo and Selection ViewModels
